Create useStyles hook
