+++ src/components/Client/Navbar.jsx
@@ -1,13 +1,6 @@
 import { useState, useEffect } from 'react'
 import { useLocation } from 'react-router-dom'
 import {
-  CLIENT_DASHBOARD_LINK,
-  CLIENT_DOCTORS_LINK,
-  CLIENT_PATIENTS_LINK,
-  CLIENT_DEPARTMENTS_LINK,
-} from '../../services/constants/links'
-import {
-  createStyles,
   Header,
   Container,
   Group,
@@ -18,90 +11,17 @@ import {
   Divider,
 } from '@mantine/core'
 import { useDisclosure } from '@mantine/hooks'
-import Logo from '../Logo'
 import { TbLogout } from 'react-icons/tb'
+import Logo from '../Logo'
+import { HEADER_HEIGHT } from '../../services/constants/styles'
+import useStyles from '../../services/hooks/useStyles'
+import {
+  CLIENT_DASHBOARD_LINK,
+  CLIENT_DOCTORS_LINK,
+  CLIENT_PATIENTS_LINK,
+  CLIENT_DEPARTMENTS_LINK,
+} from '../../services/constants/links'
 import { adminNavLinks } from '../../services/constants/navLinks'
-
-const HEADER_HEIGHT = 70
-
-const useStyles = createStyles((theme) => ({
-  root: {
-    position: 'relative',
-    zIndex: 1,
-  },
-
-  dropdown: {
-    position: 'absolute',
-    top: HEADER_HEIGHT,
-    left: 0,
-    right: 0,
-    zIndex: 0,
-    borderTopRightRadius: 0,
-    borderTopLeftRadius: 0,
-    borderTopWidth: 0,
-    overflow: 'hidden',
-
-    [theme.fn.largerThan('sm')]: {
-      display: 'none',
-    },
-  },
-
-  header: {
-    display: 'flex',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    height: '100%',
-  },
-
-  links: {
-    [theme.fn.smallerThan('sm')]: {
-      display: 'none',
-    },
-  },
-
-  burger: {
-    [theme.fn.largerThan('sm')]: {
-      display: 'none',
-    },
-  },
-
-  link: {
-    display: 'block',
-    lineHeight: 1,
-    padding: '8px 12px',
-    borderRadius: theme.radius.sm,
-    textDecoration: 'none',
-    color:
-      theme.colorScheme === 'dark'
-        ? theme.colors.dark[0]
-        : theme.colors.gray[7],
-    fontSize: theme.fontSizes.sm,
-    fontWeight: 500,
-
-    '&:hover': {
-      backgroundColor:
-        theme.colorScheme === 'dark'
-          ? theme.colors.dark[6]
-          : theme.colors.gray[0],
-    },
-
-    [theme.fn.smallerThan('sm')]: {
-      borderRadius: 0,
-      padding: theme.spacing.md,
-    },
-  },
-
-  linkActive: {
-    '&, &:hover': {
-      backgroundColor: theme.fn.variant({
-        variant: 'light',
-        color: theme.primaryColor,
-      }).background,
-      color: theme.fn.variant({ variant: 'light', color: theme.primaryColor })
-        .color,
-    },
-  },
-}))
 
 const Navbar = () => {
   const [opened, { toggle, close }] = useDisclosure(false)
