+++ src/services/constants/styles.js
@@ -1,0 +1,1 @@
+export const HEADER_HEIGHT = 70
